Refactor projectController.js
