+++ controller/ProjectController.js
@@ -1,4 +1,5 @@
 const projects = require("../data");
+const { getPostData } = require("../utlils");
 
 function sendResponse(res, status, data) {
   res.writeHead(status, { "Content-Type": "application/json" });
@@ -6,33 +7,30 @@ function sendResponse(res, status, data) {
 }
 
 function getAllProjects(req, res) {
-  console.log("getting all projects");
+  if (!projects.length) {
+    return sendResponse(res, 404, { error: "Projects not found!" });
+  }
   sendResponse(res, 200, projects);
 }
 
 function getTop3ProjectsByRevenue(req, res) {
-  console.log("getting top projects");
   const topPerformProjects = projects
     .filter((project) => project.revenue > 0)
     .sort((a, b) => b.revenue - a.revenue)
     .slice(0, 3);
 
-  sendResponse(res, 200, topPerformProjects);
-}
-
-function getTopProjectsCountByRevenue(req, res, count) {
-  console.log("getting top projects");
-  const topPerformProjects = projects
-    .filter((project) => project.revenue > 0)
-    .sort((a, b) => b.revenue - a.revenue)
-    .slice(0, count);
+  if (!topPerformProjects.length) {
+    return sendResponse(res, 404, { error: "Top performed projects not found!" });
+  }
 
   sendResponse(res, 200, topPerformProjects);
 }
 
 function getCompletedProjects(req, res) {
-  console.log("getting completed projects");
   const completedProjects = projects.filter((project) => project.isCompleted);
+
+  if (!completedProjects) 
+    return sendResponse(res, 404, { error: "No completed projects" });
 
   sendResponse(res, 200, completedProjects);
 }
@@ -40,56 +38,42 @@ function getCompletedProjects(req, res) {
 async function createProject(req, res) {
   try {
     const body = await getPostData(req);
-
     const newProject = JSON.parse(body);
 
-    if (projects.length == 0) {
+    if (projects.length == 0) 
       newProject.id = 1;
-    } else {
+    else 
       newProject.id = projects[projects.length - 1].id + 1;
-    }
-
-    console.log("Adding Project", newProject);
 
     projects.push(newProject);
-
     sendResponse(res, 200, JSON.stringify(newProject));
-  } catch (error) {
-    sendResponse(
-      res,
-      500,
-      JSON.stringify({
-        error: "Internal Server Error",
-        message:
-          "An error occurred while trying to add the new project to the database.",
-      })
-    );
+  } 
+  catch (ex) {
+    const err = { 
+      error: "Internal Server Error", 
+      message: "An error occurred while trying to add the new project to the database."
+    };
+    sendResponse( res, 500, JSON.stringify(err));
   }
 }
 
 function deleteProject(req, res, pid) {
   try {
-    console.log("Trying to delete project", pid);
     const index = projects.findIndex((project) => project.id === pid);
-
-    console.log("Index", index);
-
     projects.splice(index, 1);
 
     if (index === -1) {
-      sendResponse(
-        res,
-        404,
-        JSON.stringify({
-          error: "Not Found",
-          message: "The project with the specified ID was not found",
-        })
-      );
-      return;
-    } else {
+      const err = {
+        error: "Not Found", 
+        message: "The project with the specified ID was not found"
+      }
+      return sendResponse( res, 404, JSON.stringify(err));
+    } 
+    else {
       sendResponse(res, 200, JSON.stringify({ message: pid }));
     }
-  } catch (error) {
+  } 
+  catch (error) {
     console.log(error);
   }
 }
@@ -97,27 +81,7 @@ function deleteProject(req, res, pid) {
 module.exports = {
   getAllProjects,
   getTop3ProjectsByRevenue,
-  getTopProjectsCountByRevenue,
   getCompletedProjects,
   deleteProject,
   createProject,
 };
-
-// source : https://github.com/bradtraversy/vanilla-node-rest-api/blob/master/utils.js
-function getPostData(req) {
-  return new Promise((resolve, reject) => {
-    try {
-      let body = "";
-
-      req.on("data", (chunk) => {
-        body += chunk.toString();
-      });
-
-      req.on("end", () => {
-        resolve(body);
-      });
-    } catch (error) {
-      reject(error);
-    }
-  });
-}
