+++ utlils.js
@@ -17,4 +17,22 @@ const handleCors = (req, res) => {
   }
 }
 
-module.exports = handleCors;
+function getPostData(req) {
+  return new Promise((resolve, reject) => {
+    try {
+      let body = "";
+
+      req.on("data", (chunk) => {
+        body += chunk.toString();
+      });
+
+      req.on("end", () => {
+        resolve(body);
+      });
+    } catch (error) {
+      reject(error);
+    }
+  });
+}
+
+module.exports = { handleCors, getPostData };
